refactor(frontend/modules): Clean-up imports in Angular modules
