+++ frontend/src/app/modules/auth/auth.module.ts
@@ -1,6 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { MatButtonModule, MatDialogModule } from '@angular/material';
 
 import { SharedModule } from '@shared/shared.module';
 
@@ -9,7 +7,7 @@ import { AuthService } from './auth.service';
 
 @NgModule({
   declarations: [LoginComponent],
-  imports: [CommonModule, MatButtonModule, MatDialogModule, SharedModule],
+  imports: [SharedModule],
   exports: [LoginComponent],
   providers: [AuthService],
 })
