+++ frontend/src/app/modules/intervention-form/intervention-form.module.ts
@@ -1,30 +1,18 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
-import {
-  MatButtonModule,
-  MatCardModule,
-  MatFormFieldModule,
-  MatInputModule,
-  MatProgressSpinnerModule,
-  MatSelectModule,
-} from '@angular/material';
+import { MatCardModule, MatFormFieldModule, MatInputModule, MatSelectModule } from '@angular/material';
 
-import { InterventionFormComponent } from './form/form.component';
 import { SharedModule } from '@shared/shared.module';
 
-const MATERIAL_IMPORTS = [
-  MatButtonModule,
-  MatCardModule,
-  MatFormFieldModule,
-  MatInputModule,
-  MatSelectModule,
-  MatProgressSpinnerModule,
-];
+import { InterventionFormComponent } from './form/form.component';
+import { FileUploaderComponent } from './file-uploader/file-uploader.component';
+import { FileThumbComponent } from './file-uploader/file-thumb/file-thumb.component';
+
+const MATERIAL_IMPORTS = [MatCardModule, MatFormFieldModule, MatInputModule, MatSelectModule];
 
 @NgModule({
-  declarations: [InterventionFormComponent],
-  imports: [CommonModule, ReactiveFormsModule, ...MATERIAL_IMPORTS, SharedModule],
+  imports: [ReactiveFormsModule, SharedModule, ...MATERIAL_IMPORTS],
+  declarations: [InterventionFormComponent, FileUploaderComponent, FileThumbComponent],
   exports: [InterventionFormComponent],
 })
 export class InterventionFormModule {}
