+++ frontend/src/app/modules/nav/nav.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { MatButtonModule, MatIconModule, MatListModule, MatSidenavModule, MatToolbarModule } from '@angular/material';
+import { MatListModule, MatSidenavModule, MatToolbarModule } from '@angular/material';
 import { RouterModule } from '@angular/router';
 import { SharedModule } from '@shared/shared.module';
 
@@ -10,16 +9,7 @@ import { MobileNavComponent } from './mobile-nav/mobile-nav.component';
 
 @NgModule({
   declarations: [NavComponent, DesktopNavComponent, MobileNavComponent],
-  imports: [
-    CommonModule,
-    MatButtonModule,
-    MatIconModule,
-    MatListModule,
-    MatSidenavModule,
-    MatToolbarModule,
-    RouterModule,
-    SharedModule,
-  ],
+  imports: [MatListModule, MatSidenavModule, MatToolbarModule, RouterModule, SharedModule],
   exports: [NavComponent, MobileNavComponent],
 })
 export class NavModule {}
