+++ frontend/src/app/modules/public-form/public-form.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MatSnackBarModule } from '@angular/material';
 
 import { SharedModule } from '@shared/shared.module';
 import { InterventionFormModule } from '@interventionForm/intervention-form.module';
@@ -8,7 +7,7 @@ import { PublicFormComponent } from './public-form/public-form.component';
 
 @NgModule({
   declarations: [PublicFormComponent],
-  imports: [CommonModule, InterventionFormModule, MatSnackBarModule, SharedModule],
+  imports: [CommonModule, InterventionFormModule, SharedModule],
   exports: [PublicFormComponent],
 })
 export class PublicFormModule {}
